changed all decimals to float

+++ src/models/nfo.py
@@ -13,16 +13,16 @@ class NFO(db.Model):
     # order detail
     action = db.Column(db.String, nullable=False, server_default="buy")
     quantity = db.Column(db.Integer, default=25)
-    entry_price = db.Column(db.DECIMAL, nullable=False)
-    exit_price = db.Column(db.DECIMAL, nullable=True)
-    profit = db.Column(db.Integer, nullable=True)
+    entry_price = db.Column(db.Float, nullable=False)
+    exit_price = db.Column(db.Float, nullable=True)
+    profit = db.Column(db.Float, nullable=True)
 
     # option specific field
     strike = db.Column(db.Integer, nullable=True)
     option_type = db.Column(db.String, nullable=True)
 
     # future specifi field
-    future_price = db.Column(db.DECIMAL, nullable=True)
+    future_price = db.Column(db.Float, nullable=True)
 
     # strategy details
     strategy = db.Column(db.String, nullable=False)
